Only copy files for CDAP packages

+++ package/scripts/ambari_helpers.py
@@ -10,8 +10,9 @@ def package(name):
   import params
   Execute(params.package_mgr + ' install -y ' + name, user='root')
   # Replace service script
-  Execute('rm -f /opt/cdap/' + name[5:] + '/bin/service')
-  Execute('cp ' + params.files_dir + 'service /opt/cdap/' + name[5:] + '/bin/service')
+  if name[:4] == 'cdap':
+    Execute('rm -f /opt/cdap/' + name[5:] + '/bin/service')
+    Execute('cp ' + params.files_dir + 'service /opt/cdap/' + name[5:] + '/bin/service')
 
 def add_repo(source, dest):
   import params
